[Refactor]PlaySerivce해체
