+++ src/Main.java
@@ -1,6 +1,11 @@
+
 public class Main {
     public static void main(String[] args) {
-        PlayService playService = new PlayService();
-        playService.play();
+        InputData inputData = new InputData();
+        CarManager carManager = new CarManager(inputData.getCarNames());
+
+        Race race = new Race(inputData.getAttempts(), carManager.getCarList());
+        race.play();
     }
+
 }
